Use floats in addition and subtraction and use library function in calculation

+++ adder/src/lib.rs
@@ -6,7 +6,7 @@ use bindings::exports::docs::adder::add::Guest;
 struct Component;
 
 impl Guest for Component {
-    fn add(a: i32, b: i32) -> i32 {
+    fn add(a: f32, b: f32) -> f32 {
         return a + b;
     }
 }
